Refactor rain to weather module
