+++ myko/src/lib/visualisation/sketch.js
@@ -10,36 +10,7 @@ import {
 import { flowfieldDraw, flowfieldSetup } from './flowfield';
 //import p5Svelte from 'p5-svelte';
 //import { linear } from 'svelte/easing';
-
-class Drop {
-  constructor(p5) {
-    this.p5 = p5;
-
-    this.x = this.p5.random(this.p5.width * 0.5, this.p5.width + 50);
-    this.y = this.p5.random(500, 250);
-    this.z = this.p5.random(0, 30);
-    this.w = this.p5.random(3);
-    this.h = this.p5.map(this.z, 0, 20, 10, 20);
-    this.vel = this.p5.map(this.z, 0, 3, 2, 3);
-    this.wind = 3;
-  }
-
-  update(windForce) {
-    this.wind = windForce;
-    this.y += this.vel;
-    this.x -= this.wind;
-  }
-
-  draw() {
-    this.p5.noStroke();
-    this.p5.rotate(this.p5.PI / this.wind);
-    let c = this.p5.color(180, 11, 86, 50);
-    //c = color(255); snow
-    this.p5.fill(c);
-    //circle(this.x, this.y, this.w, this.h); snow
-    this.p5.rect(this.x, this.y, this.w, this.h);
-  }
-}
+import { makeWeather } from './weather';
 
 let canvas, xtraCnvs, xtraCnvs2;
 let addedThings = [],
@@ -52,11 +23,6 @@ let diys = [];
 let planes = [];
 let cranes = [];
 let thoughts = [];
-let rain = [];
-let heavy;
-let currentTime;
-let lastTime;
-let timer = 0;
 
 export function preload(p5) {
   cloud = p5.loadImage('cloud0.png');
@@ -87,9 +53,6 @@ export async function setup(p5) {
   p5.imageMode(p5.CENTER);
   p5.rectMode(p5.CENTER);
   p5.pixelDensity(1);
-
-  generateDrops(5, 30, p5);
-  lastTime = p5.millis();
 
   xtraCnvs = p5.createGraphics(p5.windowWidth, p5.windowHeight - 50);
   xtraCnvs.imageMode[xtraCnvs.CENTER];
@@ -142,25 +105,7 @@ export function draw(p5) {
     atm.edge();
   }
 
-  currentTime = p5.millis();
-  timer += currentTime - lastTime;
-  lastTime = currentTime;
-
-  if (timer > 300) {
-    heavy = p5.random(0, 40);
-    generateDrops(heavy / 2, heavy * 2, p5);
-
-    timer = 0;
-  }
-
-  for (let i = 0; i < rain.length; i++) {
-    if (rain[i].y > p5.height) {
-      rain.splice(i, 1);
-    } else {
-      rain[i].update(1);
-      rain[i].draw();
-    }
-  }
+  makeWeather(p5);
 }
 
 async function fetchActivityLog(p5) {
@@ -276,9 +221,3 @@ function showMoving(p5, nr, type, typeName, varySize, location1, location2, rota
   }
 }
 
-function generateDrops(min, max, p5) {
-  for (let i = min; i < max; i++) {
-    let drop = new Drop(p5);
-    rain.push(drop);
-  }
-}
